Use restrict-password-params branch of devise_security_extension
We need this gem to be using strong params in the PasswordExpiredController.
At the same time we're moving to only permit the 3 password params
which update_with_password (in Devise itself) needs, so that it isn't
possible to set any other attributes on the user when updating the
password.

We were already using a branch of a fork of this gem. Both of the
changes on this branch should be PRed upstream.